fix packages dep
