+++ reporting-service/pom.xml
@@ -265,6 +265,7 @@
                     </dependency>
                 </dependencies>
                 <configuration>
+                    <!--suppress UnresolvedMavenProperty -->
                     <argLine>${argLine}</argLine>
                 </configuration>
             </plugin>
@@ -289,7 +290,7 @@
                                         <extraarg>-verbose</extraarg>
                                         <extraarg>-xjc-verbose</extraarg>
                                         <extraarg>-p</extraarg>
-                                        <extraarg>it.gov.pagopa.hubpa.servicewsdl</extraarg>
+                                        <extraarg>it.gov.pagopa.reporting.servicewsdl</extraarg>
                                     </extraargs>
                                 </wsdlOption>
                             </wsdlOptions>
