+++ reporting-service/src/main/java/it/gov/pagopa/reporting/models/BooleanResponseModel.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.hubpa.models;
+package it.gov.pagopa.reporting.models;
 
 import lombok.Data;
